Clean up errors raised in debug mode
